auditor: add a config option to enable wvfs path auditor caching
Summary:
The reason caching is disabled seems to be that symlinks can change where
paths points to, while some properties of that path are cached.

Sometimes we know that symlinks are not used in the repo, so we'd like hg
to let us shoot ourselves in the foot.

The motivation here is that I have a specially crafted rebase operation,
which makes on the order of 400000 stat-like syscalls without caching and
on the order of 6000 with caching. This can save ~20s on Windows.

Note that this setting is disabled by default so we can't break anything.
FWIW, we already have an unsafe config option, which allows us to opt into the "we don't do crazy stuff with symlinks" behavior, it's called `unsafe.filtersuspectsymlink`, setting it to false speeds up Windows `status` and we do so for `ovrsource` on Windows.

Reviewed By: quark-zju

Differential Revision: D8297129

fbshipit-source-id: ec751d18b50b042f28f68c9e1a3634ab94a4e7c3