bundle2-resolver: reorganize the code so that returned Error's have proper context attached
Summary:
This change screams for using something like "await", but because we don't have it yet we need to reorganize this code into a chain of `and_then`.
Thanks to that if we attach a context to a future inside `and_then` the context will be properly reported when an error happens inside that future.

This is a bit of a simplification, but the previous code used the equivalent of `map` rather than `and_then` and this resulted in Errors returned from within f.e. `upload_changesets` to have a context of "While resolving B2xInfinitepushBookmarks" attached which was confusing.

Reviewed By: StanislavGlebik

Differential Revision: D8284606

fbshipit-source-id: b8a4bd63b9732caea53089f1b9088f2624a42f87