fsmonitor: add config around hashignore checking
Summary:
This fixes the issue where every merge conflict situation immediately took 30s - 1m due to the "scanning working copy" full repo scan.

This happens because sparse matcher with two working parents is different from with one working parent - It switches from returning one matcher for the gitignore to returning both parent's matchers in a unionmatcher -- even if they are identical, the hash changes, triggering a full repo can.

After discussing with quark-zju, he mentioned the fsmonitor gitignore change detection has been mostly-broken for a long time ([1], [2]), and is no longer worth maintaining. The perf costs outweigh correctness and implementation of "correctness" costs, which are esoteric (you update to master with a dirty file that master happens to have unignored -- but unignores are rare) and easily fixed by the user (saving any of the files fixes it). Lastly, quark-zju's treestate stack already does this, so this effectively brings those perf benefits early.

In the longer team, the correct fix would be tracking ignored files (including those outside sparse) in treestate, and examining them.

Wrapped in a config in case it goes wrong.

[1] https://www.mercurial-scm.org/repo/hg-committed/rev/6dea1701f170f38f931759267d2bbd3f49de92a5 (.hgignore only invallidated if the toplevel .hgignore changed, which was rare -- at Facebook usually the subdirectory ignore was changed)
[2] In `matcher.py`, the `gitignorematcher.__repr__` returns a static string.

Reviewed By: quark-zju

Differential Revision: D8337682

fbshipit-source-id: 725b3eb8fba06940fd2db0903bcecf7416e64e0d