revert: remove the special-cased code for always matchers
Summary:
Remove the special-case code path where m.always() is true.  The non-always
code path also behaves correctly with always matchers and should perform just
as efficiently.  The only two additional things it does is to customize the
matcher's `bad()` callback (which should never be invoked for matchers that
match everything) and to loop through `m.files()`, which is guaranteed to be
empty for matchers where `m.always()` is true.

Reviewed By: quark-zju

Differential Revision: D8229267

fbshipit-source-id: adb9f7cfe269febd53ff417ec0d3817833ac5338