mononoke: add SinkAsyncWrite adapter
Summary:
AsyncWrite and Sink traits have very similar interfaces. This diff creates an
adapter between them.

The primary motivation for this adapter is in Mononoke. Currently mononoke
wireproto methods that return bundles have to buffer them in memory. This have
a few drawbacks. First of all, it increases memory usage and secondly it
increases the latency.

The proper fix would to make Bundle2Encoder return a Stream instead of
BoxFuture<(), Error>. However that would require changing the whole
async-compression crate, and that's too difficult. We may want to eventually do
it, however for now SinkAsyncWrite seems like a good enough fix.

Reviewed By: farnz

Differential Revision: D8379585

fbshipit-source-id: 19af9452ba09318a7505dda44ef765e8c09b004d