mononoke: use SinkAsyncWrite in gettreepack
Summary:
Use SinkAsyncWrite that was added in the previous diff.
The main motivation is to make sure that we do not buffer bundle2 response in
memory. For more details look at the summary from the previous diff.

Note that this diff still buffers results in memory, but it prepares the code
for the next diffs, in which we'll finally remove the buffering.

Reviewed By: farnz

Differential Revision: D8379586

fbshipit-source-id: 5985ddb157f7066732ef47a04919eb1fba68c74e