commitcloud: determine sync obsmarkers at the same time as heads and bookmarks
Summary:
We get the list of obsmarkers to sync after we have pushed the commits.  In
that time, new obsmarkers could have shown up, which shouldn't be synced until
the next sync operation.

Improve this by getting the obsmarkers just after we get the heads and
bookmarks.  This still has a race, but the window is now much smaller.
It also means we can ensure we sync even when all we have is some extra
obsmarkers.

Reviewed By: DurhamG

Differential Revision: D8397739

fbshipit-source-id: e29890aa250d58bfc3c3c2a056a01d287c053605