commitcloud: update bookmarks and obsmarkers as part of the sync pull
Summary:
Avoid transient partially synced stated by updating the bookmarks and
obsmarkers as part of the pull command, if there is one.   Additionally,
prevent remotenames bookmarks from being updated in cloud-sync pulls that
happen to include public commits.  That's preferable, as usually this will only
be a subset of remote bookmarks, and this can give confusing results.

Reviewed By: DurhamG

Differential Revision: D8401117

fbshipit-source-id: 5ee646d8c7e49ea70d3ec5fd7df5c9009de7bcdd