commitcloud improve error message for access keychain from ssh sessions
Summary:
security command normally show a GUI prompt to allow you to type your password if the keychain is locked
if it is unable to do so, it returns 36 and empty stderr

it is ok that it is failed, because it happens due to security reasons,
the problem is only that stderr is empty, so the reason of failure is not clear to user.

I didn't find any option in the man of security command itself. So, I checked the status code.

Reviewed By: markbt

Differential Revision: D8418363

fbshipit-source-id: 0c8f0bf55f588822e50395c3ec8bfa8d3188ecf7