rebase: new, faster fallback logic for conflicts
Summary:
The previous logic restarted the *entire* rebase if one commit had merge conflicts or needed a mergedriver run. This was unnecessarily slow, plus the abort triggered a full repo scan.

Also, this logic is actually simpler and easier to reason about when done here, vs. `rebase()`. We can delete all the wrapping that function does after this rolls out.

This is config-gated, so we can roll it out gradually. Old config = old behavior.

Reviewed By: quark-zju

Differential Revision: D8272566

fbshipit-source-id: 9e598eea10c8a634497cd211a468dd2f58c5ff47