move the logic for committing to a separate method
Summary:
This commit moves the logic concerning creation of commit during the
graft step in pushrebase to a separate method. This allows for easily wrapping
around the commit to add custom logic. In particular, this can be used to add
extras for the commits that will be created using the pushrebase.

Reviewed By: DurhamG

Differential Revision: D8364383

fbshipit-source-id: 60f845ec3f4f0b4d3799f4385ce49c6fd78cdda2