rebase: invalidate old mergedriver preprocess during IMM restart
Summary:
This fixes a subtle bug related to mergedrivers and IMM.

If your DAG looks like this:

```
     dest (new mergedriver code)
      |
      |  A (current commit)
      |/
     base (old mergedriver code)
```

And if you use IMM to rebase `A` onto `dest`, the wrong thing can happen because parts of `base`'s mergedriver are still loaded.

This only affected certain types of driver changes -- adding or removing parameters between files in the driver, or a driver change that also mutated the working copy (e.g. renaming a file checked by `preprocess()`). And, it only occurred if you had `A` (or some other change between `base` and `dest) checked out.

So, evict the old driver files, and all is well.

The right long-term fix is to probably implement our own Python loader for the mergedriver that reads all driver files from either the `source` or `dest` changectx not `.`. But, that's a bigger change.

Reviewed By: quark-zju

Differential Revision: D8464848

fbshipit-source-id: 8f03a4c99418aa26c3890989d19dd908b026c328