localrepo: decide whether to cache wvfs auditor after reading repo configs
Summary:
Now, it is impossible to turn `wvfs` auditor caching from repo configs, as
they are read after `wvfs` is instanciated and thus don't influence this
decision.

A proper fix would be to refactor the `localrepository.__init__` method to do
these things in the right order, but for now I would like to just fix what is
currently broken.

Reviewed By: markbt

Differential Revision: D8447069

fbshipit-source-id: d45c1424a5b680c234cd804664599c2074371668