Refactor lua hooks to use co-routines
Summary:
This diff refactors the lua hooks implementation to use coroutines.
This means that if a lua hook yields a future (e.g. a call to make a network request), then the thread executing the hook can go off and do other stuff until the network result is ready at which point it can be resumed.
In an ideal world this means that Lua hooks shouldn't take much CPU unless someone decides to compute large Fibonacci numbers with them or run a bitcoin miner, but hopefully that won't occur. So, for now I have removed the code that spawns the hooks on a thread pool but we can consider re-instating that in the future if necessary.
The implementation of this was quite challenging, as the lua coroutine needs to own its Lua instance as it can't reference an existing Lua instance owned outside because of lifetime rules for Futures. Unfortunately the existing lua::get method which retrieves the hook function (mutably) borrows self (lua) making it impossible to use for our purposes. (Most probably the lua library wasn't built with future use cases in mind).
To work around that, we have to get the reference to the hook function using direct lua ffi (this is more or less what the get method does internally in lua anyway)

Reviewed By: sunshowers

Differential Revision: D8382301

fbshipit-source-id: aa53438e350e249a56413189580f53b5b8f19f40