Refactor hook manager to take changeset_id and lookup changeset internally
Summary: This diff refactors the hook manager to run hooks based on a changeset_id not the actual changeset. The actual changeset is looked up by the hook manager lazily when running a hook. This is needed to make it work well for AsyncMemo as we don't want the changeset cached in the key.

Reviewed By: StanislavGlebik

Differential Revision: D8422442

fbshipit-source-id: 40bc89124942e05c5aaeb2b4ee00215afd816642