add test for pulling public commits by revision
Summary:
Add a test for pulling public treeonly commits by their revision number.
Unlike a normal pull, this doesn't use the fastpath, and attempts to
recalculate some information.  It tries to calculate too much, and as a result
ends up diffing each flat manifest pair between the client head and the
requested revision.  For large numbers of commits, this is really slow.

Reviewed By: DurhamG

Differential Revision: D8446138

fbshipit-source-id: 5e2cbb7e710282b094d4fe93b1f7a37afaf7ab3d