rebase: manually remove rebasestate
Summary:
This is a quick bug fix for the situation xixixao discovered:

You rebase commit B onto C while having A checked out, and the rebase creates merge conflicts.

In that case, the command aborts complaining of the uncommitted changes. But, the rebasestate isn't properly removed. The user might logically run `hg rebase --abort` to clear it, but that does an
`update` with `force=True` which can cause work to be lost.

This is just a quick bug fix. The right solution is to make an inmemory `mergestate` and not write it at all if rebasing in-memory.

Reviewed By: DurhamG

Differential Revision: D8517289

fbshipit-source-id: a62c99ccb2de1be102f9c23cd737794184c4fde8