Add ability to set timeouts on Futures
Summary: This diff adds the methods `timeout()` and `on_timeout()` to the `FuturesExt` trait. The former sets a time limit for the execution of a `Future` -- effectively shorthand for `tokio_timer::Deadline`. The latter takes a callback (which may itself return a `Future`) which is called if the wrapped `Future` times out; this is useful for doing things like logging the state of the program to debug why things are timing out.

Reviewed By: StanislavGlebik

Differential Revision: D8508008

fbshipit-source-id: 6d085c35b68d1cf5a24446be8af77eb30028a7db