store: workaround long non-fncache file names
Summary:
hg filename encoding will double the length of upper-case paths. That could
be problematic if there is a long file name with many upper-case characters.
Workaround it by detecting the long name case, and keep upper-case
unchanged.

This unblocks a production issue where a long upper-case file name was
committed to the source of truth and the hg mirror cannot import it. But it
makes Windows hosting less "correct".

Reviewed By: DurhamG

Differential Revision: D8527475

fbshipit-source-id: da25fcce7e3173769788bd164629c28466158cd2