treemanifest: fix backfilling when there are local-only public heads
Summary:
Flat manifest backfilling (which is used when reverting from treeonly
mode) was failing when the client had public heads that didn't exist on the
server (for instance, if they pulled from svn directly or if they had old
commits that had been stripped from the server). This was breaking the pull.

Let's fix it so we never send unknown heads to the server during the backfill
pull.

Reviewed By: phillco

Differential Revision: D8526027

fbshipit-source-id: 8d5bf67d2022dbc24facd0b96e001c8c81a0a2da