bundle2-resolver: pass only necessary entries for persisting a changeset
Summary:
Prior to this diff when we constructed list of child entries to be passed to CreateChangeset we have been walking the tree of manifests and appending everything that we have found to the child entries list, which often include the entries of the commit parent if it is pushed in the same unbundle command.

After this diff we append only the things that are connected to the diff to be created filtering out the ones that will be handled by parent creation.

This code assumes that the client doesn't send us entries that are part of parent that were pushed in a separate unbundle command. If we wouldn't make that assumption then we would need to consult the blobstore in order to filter out those entries.

Reviewed By: StanislavGlebik

Differential Revision: D8547867

fbshipit-source-id: ac72f1b6066c15b1fea029e0224fe0ebf9a2db59