treemanifest: consider a tree remote from a client if it doesn't exist locally
Summary:
In the recent refactor we changed the definition of "is this tree
local" on the server to also include if the tree didn't exist in the store,
since then it might be an ondemand conversion from a flat manifest in a bundle.
This caused issues on peer-to-peer pulls, since most trees don't exist locally.

Let's fix it by only considering a missing tree to be a local tree if we're on
the server.

Reviewed By: quark-zju

Differential Revision: D8586779

fbshipit-source-id: ea8c8ee3e2a4dc054fae9caa122cfc733df1ab81