revset: increase weight of public revsetpredicate
Summary:
The `public()` revsetpredicate has the default weight of 1, which means
expressions like `::x & public()` are optimized to `('andsmally', ::x, public())`,
which will actually result in iterating over all commits in the repo.

Increase the weight to 3 to ensure it's optimized in a way that favours
following ancestry links and then checking the phase.

Reviewed By: ikostia

Differential Revision: D8549645

fbshipit-source-id: f58750ecb1918079b7e350ce27b7a61771489233