add a timeout in the CLI for unmount operations
Summary:
This updates the CLI code to put a timeout on `unmount()` thrift calls to
edenfs.  We've had a few reports of situations where the unmount thrift call
has hung while edenfs was waiting for all inodes to be released.

Eventually we should fix the C++ code to put a hard timeout on the operation,
(and forcibly remove the mount point from its list in EdenServer even if it is
still waiting for some inodes to be released).  However as an interim solution
timing out on the CLI is better than hanging forever.  In particular the old
behavior made `eden rm` unable to clean up some mount points.

Reviewed By: pkaush

Differential Revision: D8558767

fbshipit-source-id: 942e5c7bcc147229f040f377ce59fcc2565276e0